+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.17.1
 
 - Fixed AaronUtil#getGridSpiralPos
+- Fixed MenuWithInventory#quickMoveStack when machine slots are added after player inventory slots
 
 # 1.17.0